Open-BFME5: BFME renders on Direct3D 9, and dx8indexbuffer.cpp is now complete
The DX8IndexBufferClass constructor at 0x00138980 said so three separate ways,
and all three had to be true together for its 276 bytes to match:

  * CreateIndexBuffer is called at device vtable +6Ch. D3D8 puts it at +60h;
    D3D9 puts it at +6Ch, because D3D9 adds GetSwapChain,
    GetNumberOfSwapChains and SetDialogBoxMode ahead of it.
  * That call pushes six arguments, not five. The extra `push 0` is D3D9's
    pSharedHandle.
  * DX8Caps::Support_TnL reads `[ecx+138h]`. MaxDisplayWidth, MaxDisplayHeight
    and Caps put that flag at 8 + sizeof(D3DCAPS), so retail's structure is 304
    bytes -- D3DCAPS9's size, where reference/shims/d3d8_shim_validated.h stops
    at 212.

None of that is inference from a name: the body byte-matches only when the
device interface has D3D9's slot layout and D3D9's signature. The shim change
is behind BFME_D3D9_DEVICE_ABI so no other unit moves until it opts in, and the
92 caps bytes are a named reserved tail rather than invented field names --
nothing in this tree reads them.

Two smaller BFME differences from the same body: the count is `unsigned` all
the way down, including IndexBufferClass's own constructor, where an
`unsigned short` parameter costs a movzx in every derived constructor; and the
buffer-creation retry path drops Zero Hour's Invalidate_Old_Unused_Textures
sweep, keeping only WW3D::_Invalidate_Mesh_Cache.

With the layout and the ABI both right, place_bodies found four more on the
next pass, and the file is done:

  0x001386C0   69  IndexBufferClass::IndexBufferClass
  0x00138980  276  DX8IndexBufferClass::DX8IndexBufferClass
  0x00138E00  328  DynamicIBAccessClass::Allocate_DX8_Dynamic_Buffer
  0x001390C0  145  IndexBufferClass::Copy(unsigned int*)
  0x00139160  145  IndexBufferClass::Copy(unsigned short*)

dx8indexbuffer.cpp went from 5 matched functions this morning to 23, with no
present-unmatched markers left in it.

The include order needed one more note. dx8wrapper.h includes dx8caps.h from
its own directory, and a quoted include searches the including file's directory
before the -I list, so the reference copy won the include guard before the shim
was ever consulted. Pulling dx8caps.h in first, by -I, from the .cpp fixes it.

+++ reference/shims/indexbuffercount/dx8indexbuffer.h
@@ -80,7 +80,9 @@ class IndexBufferClass : public W3DMPO, public RefCountClass
 protected:
 	virtual ~IndexBufferClass();
 public:
-	IndexBufferClass(unsigned type, unsigned short index_count);
+	// The count is `unsigned` all the way down in BFME -- an unsigned short
+	// parameter here costs a movzx in every derived constructor.
+	IndexBufferClass(unsigned type, unsigned index_count);
 
 	void Copy(unsigned int* indices,unsigned start_index,unsigned index_count);
 	void Copy(unsigned short* indices,unsigned start_index,unsigned index_count);
@@ -205,7 +207,10 @@ public:
 		USAGE_NPATCHES=4
 	};
 
-	DX8IndexBufferClass(unsigned short index_count,UsageType usage=USAGE_DEFAULT);
+	// BFME's constructor takes the count as `unsigned`: retail reads the
+	// argument with a plain 32-bit load where an unsigned short parameter
+	// forces a movzx.
+	DX8IndexBufferClass(unsigned index_count,UsageType usage=USAGE_DEFAULT);
 	~DX8IndexBufferClass();
 
 	void Copy(unsigned int* indices,unsigned start_index,unsigned index_count);
